can now read/write file-like objects; a Writer to be used as a context manager

+++ msl/io/writers/json_.py
@@ -7,6 +7,7 @@ Writer for a JSON_ file format. The corresponding :class:`~msl.io.base_io.Reader
 import os
 import json
 import codecs
+from io import BufferedIOBase
 
 import numpy as np
 
@@ -46,9 +47,9 @@ class JSONWriter(Writer):
 
         Parameters
         ----------
-        url : :class:`str`, optional
-            The name of the file to write to. If :data:`None` then uses the value of
-            `url` that was specified when :class:`JSONWriter` was created.
+        url : :term:`path-like <path-like object>` or :term:`file-like <file object>`, optional
+            The file to write the `root` to. If :data:`None` then uses the value of
+            `url` that was specified when :class:`JSONWriter` was instantiated.
         root : :class:`~msl.io.base_io.Root`, optional
             Write `root` in JSON_ format. If :data:`None` then write the
             :class:`~msl.io.group.Group`\\s and :class:`~msl.io.dataset.Dataset`\\s
@@ -60,14 +61,15 @@ class JSONWriter(Writer):
             `json.dump <https://docs.python.org/3/library/json.html#json.dump>`_.
             The default indentation is 2.
         """
-        url = self.url if url is None else url
+        if url is None:
+            url = self.url
         if not url:
             raise ValueError('You must specify a url to write the file to')
 
         if root is None:
             root = self
         elif not isinstance(root, Root):
-            raise TypeError('the root parameter must be a Root object')
+            raise TypeError('The root parameter must be a Root object')
 
         def add_dataset(d, dataset):
             if dataset.dtype.fields:
@@ -118,9 +120,11 @@ class JSONWriter(Writer):
             # (on Windows) when the standard open function use used
             opener = open
 
+        is_file_like = hasattr(url, 'write')
+
         if not open_kwargs['mode']:
             open_kwargs['mode'] = 'w'
-            if os.path.isfile(url):
+            if not is_file_like and os.path.isfile(url):
                 raise IOError("A {!r} file already exists.\n"
                               "Specify mode='w' if you want to overwrite it.".format(url))
 
@@ -130,9 +134,23 @@ class JSONWriter(Writer):
             kwargs['cls'] = _NumpyEncoder
             json.encoder._make_iterencode = _make_iterencode
 
-        with opener(url, **open_kwargs) as fp:
-            fp.write('#File created with: MSL {} version 1.0\n'.format(self.__class__.__name__))
-            json.dump(dict_, fp, **kwargs)
+        header = '#File created with: MSL {} version 1.0\n'.format(self.__class__.__name__)
+
+        if is_file_like:
+            if isinstance(url, BufferedIOBase):  # a bytes-like object is required
+                encoding = open_kwargs['encoding']
+                url.write(header.encode(encoding))
+                url.write(json.dumps(dict_, **kwargs).encode(encoding))
+            elif IS_PYTHON2:
+                url.write(unicode(header))
+                url.write(unicode(json.dumps(dict_, **kwargs)))
+            else:
+                url.write(header)
+                json.dump(dict_, url, **kwargs)
+        else:
+            with opener(url, **open_kwargs) as fp:
+                fp.write(header)
+                json.dump(dict_, fp, **kwargs)
 
         if kwargs['cls'] is _NumpyEncoder:
             json.encoder._make_iterencode = _original_make_iterencode
